Revert "Revert "Revert "Git clone the user repo with '--depth=10 --no-single-branch' so we have the clone has the previous 10 commits for every branch that the project has (instead of cloning all the history for every branch)"""
This reverts commit 0c4f40d5a12990b7f6cb25fdb29bdb47b1cd6c1d.

We realized that there is no point in doing a shallow clone of the
mirror repo as it is not part of the container but of the cache dir. We
do need a complete bare clone in /var/lib/git and it is the build which
should do the shallow clone.